refactor(frontend): shared blacklist presentation and detail fixtures
Blacklist amount presentation was formatted independently in the table and
the detail event feed, FlowChart re-implemented what the ChartFigure
primitive already provides, and the stablecoin-detail test input factory was
duplicated across two consumers.

Also deletes two test-only islands: src/lib/mint-burn-timeframes.ts and the
src/lib/urls.ts compatibility shim, each reachable only from its own test,
plus the three timeframe constants in src/lib/constants.ts that the
timeframe module was the last consumer of (DAY_HOURS stays; FlowChart uses
it).

FH-04, the unwritable motion override, is intentionally left in place: the
recon report flagged it as needing a product decision rather than a
deletion.
