--- src/lib/__tests__/urls.test.ts
@@ -1,9 +1,0 @@
-import { describe, expect, it } from "vitest";
-import { buildStablecoinUrl as buildCompatibilityUrl } from "@/lib/urls";
-import { buildStablecoinUrl } from "@shared/lib/urls";
-
-describe("stablecoin URL compatibility entrypoint", () => {
-  it("re-exports the runtime-neutral canonical builder", () => {
-    expect(buildCompatibilityUrl).toBe(buildStablecoinUrl);
-  });
-});
