--- src/lib/urls.ts
@@ -1,3 +1,0 @@
-// Compatibility entrypoint for downstream consumers outside the frontend.
-// New code should import the runtime-neutral owner from @shared/lib/urls.
-export { buildStablecoinUrl } from "@shared/lib/urls";
